Cleanup paginated_embeds impl
